edit black pieces out of board position

+++ common/chessPosition.hpp
@@ -2,8 +2,8 @@
 #define CHESS_POSITION_HPP
 
 struct ChessPosition {
-    int x; // 0-7 对应 a-h
-    int y; // 0-7 对应 1-8
+    float x; // 0-7 对应 a-h
+    float y; // 0-7 对应 1-8
 };
 
 #endif // CHESS_POSITION_HPP
